refactor: extract transcription guide hook
